+++ packages/components/src/components/Tabs/Tabs.stories.tsx
@@ -5,6 +5,8 @@ import { action } from '@storybook/addon-actions';
 import { type Meta, type StoryObj } from '@storybook/react';
 import { useState } from 'react';
 
+import { Badge } from '../Badge/Badge';
+
 import { Tabs, TabsContent, TabsRoot, TabsTab, TabsTrigger } from './Tabs';
 
 type Story = StoryObj<typeof meta>;
@@ -214,6 +216,44 @@ export const WithDecorator: Story = {
     ),
 };
 
+export const WithBadge: Story = {
+    render: (args) => (
+        <Tabs.Root {...args}>
+            <Tabs.Tab value="first">
+                <Tabs.Trigger>
+                    First Tab <Badge variant="highlight">Beta</Badge>
+                </Tabs.Trigger>
+                <Tabs.Content>
+                    <b>First Content</b>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat.
+                </Tabs.Content>
+            </Tabs.Tab>
+            <Tabs.Tab value="second">
+                <Tabs.Trigger>Second Tab</Tabs.Trigger>
+                <Tabs.Content>
+                    <b>Second Content</b>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat.
+                </Tabs.Content>
+            </Tabs.Tab>
+            <Tabs.Tab value="third">
+                <Tabs.Trigger>
+                    Third Tab <Badge variant="highlight">Beta</Badge>
+                </Tabs.Trigger>
+                <Tabs.Content>
+                    <b>Third Content</b>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat.
+                </Tabs.Content>
+            </Tabs.Tab>
+        </Tabs.Root>
+    ),
+};
+
 export const WithNoPadding: Story = {
     args: {
         padding: 'comfortable',
@@ -377,6 +417,78 @@ export const WithSpaciousPadding: Story = {
             </Tabs.Tab>
             <Tabs.Tab value="third">
                 <Tabs.Trigger>Third Tab</Tabs.Trigger>
+                <Tabs.Content>
+                    <b>Third Content</b>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat.
+                </Tabs.Content>
+            </Tabs.Tab>
+        </Tabs.Root>
+    ),
+};
+
+export const WithPillVariant: Story = {
+    render: (args) => (
+        <Tabs.Root variant="pill" {...args}>
+            <Tabs.Tab value="first">
+                <Tabs.Trigger>First Tab</Tabs.Trigger>
+                <Tabs.Content>
+                    <b>First Content</b>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat.
+                </Tabs.Content>
+            </Tabs.Tab>
+            <Tabs.Tab value="second">
+                <Tabs.Trigger>Second Tab</Tabs.Trigger>
+                <Tabs.Content>
+                    <b>Second Content</b>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat.
+                </Tabs.Content>
+            </Tabs.Tab>
+            <Tabs.Tab value="third">
+                <Tabs.Trigger>Third Tab</Tabs.Trigger>
+                <Tabs.Content>
+                    <b>Third Content</b>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat.
+                </Tabs.Content>
+            </Tabs.Tab>
+        </Tabs.Root>
+    ),
+};
+
+export const PillVariantWithBadge: Story = {
+    render: (args) => (
+        <Tabs.Root variant="pill" {...args}>
+            <Tabs.Tab value="first">
+                <Tabs.Trigger>
+                    First Tab <Badge variant="highlight">Beta</Badge>
+                </Tabs.Trigger>
+                <Tabs.Content>
+                    <b>First Content</b>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat.
+                </Tabs.Content>
+            </Tabs.Tab>
+            <Tabs.Tab value="second">
+                <Tabs.Trigger>Second Tab</Tabs.Trigger>
+                <Tabs.Content>
+                    <b>Second Content</b>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat.
+                </Tabs.Content>
+            </Tabs.Tab>
+            <Tabs.Tab value="third">
+                <Tabs.Trigger>
+                    Third Tab <Badge variant="highlight">Beta</Badge>
+                </Tabs.Trigger>
                 <Tabs.Content>
                     <b>Third Content</b>
                     Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore