+++ packages/fondue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @frontify/fondue
 
+## 12.9.6
+
+### Patch Changes
+
+- Updated dependencies [[`070c711`](https://github.com/Frontify/fondue/commit/070c7116c3015b43a0579f4214dba38e73a069e1), [`5a51c2b`](https://github.com/Frontify/fondue/commit/5a51c2b9cdc3ba89fde1e1e1d6ddd0237e105fa2), [`cc96757`](https://github.com/Frontify/fondue/commit/cc9675712f08476ae70dbdb3db5d07e1029324b4), [`6508be9`](https://github.com/Frontify/fondue/commit/6508be914ee3d6ef737a17f9fc216572374fc06d), [`d5e69a3`](https://github.com/Frontify/fondue/commit/d5e69a36561e79fe516f3ebcd02e4dede930f965), [`bd15f73`](https://github.com/Frontify/fondue/commit/bd15f73181b9e10e669244f30de7e8af77bffb10)]:
+    - @frontify/fondue-components@17.1.0
+
 ## 12.9.5
 
 ### Patch Changes